pyproject.toml
[project]
name = "humpback-acoustic-embed"
version = "0.1.0"
description = "Humpback whale acoustic embedding and clustering platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.18",
    "pyarrow>=18.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
ml = [
    "librosa>=0.10",
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "scikit-learn>=1.5",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "pytest-watch>=4.2",
    "scipy>=1.14",
]

[project.scripts]
humpback-api = "humpback.api.app:main"
humpback-worker = "humpback.workers.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/humpback"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.hatchling]
